changed database tables, added skirt calculations

+++ DAL.App.DTO/ExtraSize.cs
@@ -3,7 +3,7 @@ namespace DAL.App.DTO;
 public class ExtraSize : DomainEntityId
 {
     public string Name { get; set; } = default!;
-    public int Extra { get; set; }
+    public float Extra { get; set; }
     public Guid InstructionId { get; set; }
     public Instruction? Instruction { get; set; }
 }
